+++ src/lib/twitter-client-constants.ts
@@ -30,6 +30,8 @@ export const FALLBACK_QUERY_IDS = {
   ListMemberships: 'BlEXXdARdSeL_0KyKHHvvg',
   ListLatestTweetsTimeline: '2TemLyqrMpTeAmysdbnVqw',
   ListByRestId: 'wXzyA5vM_aVkBL9G8Vp3kw',
+  HomeTimeline: 'edseUwk9sP5Phz__9TIRnA',
+  HomeLatestTimeline: 'iOEZpOdfekFsxSlPQCQtPg',
 } as const;
 
 export type OperationName = keyof typeof FALLBACK_QUERY_IDS;